fix: ignore uppy complete event with undefined payload
With the latest uppy update, there is a bug where sometimes
the payload of complete event is undefined. This broke the event handler.
Ignoring the event in case the payload is undefined, we should get around
the issue back to the original behaviour.

+++ packages/web-pkg/src/services/uppy/uppyService.ts
@@ -276,6 +276,10 @@ export class UppyService {
       this.clearInputs()
     })
     this.uppy.on('complete', (result) => {
+      if (!result) {
+        return
+      }
+
       this.publish('uploadCompleted', result)
       result.successful.forEach((file: any) => {
         this.uppy.removeFile(file.id)
